Refactor semver comparison to formatUtils and fix i18n consistency (#66)
