+++ packages/ui-vue/src/components/views/SettingsView.Menu.vue
@@ -16,7 +16,7 @@ const value = defineModel<SettingsView>({ default: 'ai' })
 <template>
   <aside class="settings-menu">
     <TextNavigationButton v-model="value" :value="'ai'" title="AI-konfiguration" />
-    <TextNavigationButton v-model="value" :value="'extensions'" title="Tillägg" />
+    <TextNavigationButton v-model="value" :value="'extensions'" :title="$t('extensions.title')" />
     <TextNavigationButton v-model="value" :value="'localization'" title="Lokalisering" />
     <TextNavigationButton v-model="value" :value="'interface'" title="Gränssnitt" />
     <TextNavigationButton v-model="value" :value="'notifications'" title="Notiser" />
